Autowired the Pjax mapper
